# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs MlCity Project",
        "description": "API MlCity",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/event-types": {
            "get": {
                "tags": [
                    "Event-type"
                ],
                "summary": "Get all event type",
                "operationId": "a8cf2248fd70bd5747eab92fa316b898",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/getTypesId/{id}": {
            "get": {
                "tags": [
                    "Event-type"
                ],
                "summary": "Get event type by id",
                "operationId": "7ec5523eafe887e83b87fc0c193860f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/addTypeEvent/{event_id}/{type_id}": {
            "post": {
                "tags": [
                    "Event-type"
                ],
                "summary": "add type in event",
                "operationId": "3a9378f11227b180feed666f92422ed2",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "$event_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/updateTypeEvent/{event_id}/{type_id}": {
            "put": {
                "tags": [
                    "Event-type"
                ],
                "summary": "add type in event",
                "operationId": "8c6980167e8ac93ee6eecd4ac65c0224",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "$event_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/deleteTypeEvent/{event_id}/{type_id}": {
            "delete": {
                "tags": [
                    "Event-type"
                ],
                "summary": "add type in event",
                "operationId": "45c15964a4ffc7e0a3a143da369c897a",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "$event_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}